+++ client/src/components/SessionSummary/index.jsx
@@ -1,20 +1,38 @@
-function SessionSummary ({sessionData}) {
+import React from 'react';
 
-return (
-
-  sessionData.length
-    ? sessionData.map((i, j) => (
-        <tr key={j}>
-          <td>{i.targetName}</td>
-          <td>{i.commonName}</td>
-          <td>{i.sessionDate}</td>
-          <td>{i.dsoCategory}</td>
-          <td>{i.telescope}</td>
-          <td>{i.camera}</td>
+function SessionSummary({ sessionData }) {
+  return (
+    <table>
+      <thead>
+        <tr>
+          <th>Target Name</th>
+          <th>Common Name</th>
+          <th>Date</th>
+          <th>Category</th>
+          <th>Telescope</th>
+          <th>Camera</th>
         </tr>
-      ))
-    : ' '
-
-)
+      </thead>
+      <tbody>
+        {sessionData.length > 0 ? (
+          sessionData.map((item, index) => (
+            <tr key={index}>
+              <td>{item.targetName}</td>
+              <td>{item.commonName}</td>
+              <td>{item.sessionDate}</td>
+              <td>{item.dsoCategory}</td>
+              <td>{item.telescope}</td>
+              <td>{item.camera}</td>
+            </tr>
+          ))
+        ) : (
+          <tr>
+            <td colSpan="6">No data available</td>
+          </tr>
+        )}
+      </tbody>
+    </table>
+  );
 }
-export default SessionSummary
+
+export default SessionSummary;