+++ client/src/pages/Profile.jsx
@@ -136,21 +136,7 @@ const Profile = () => {
             <div id='sessions' className='col s12'>
               <h5>Sessions</h5>
               <div className='row'>
-                <table>
-                  <thead>
-                    <tr>
-                      <th>Target Name</th>
-                      <th>Common Name</th>
-                      <th>Date</th>
-                      <th>Category</th>
-                      <th>Telescope</th>
-                      <th>Camera</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    <SessionSummary sessionData={user.sessions} />
-                  </tbody>
-                </table>
+                <SessionSummary sessionData={user.sessions} />
               </div>
             </div>
           </div>