+++ CHANGELOG.md
@@ -36,3 +36,4 @@
 35. Update NavBar component to use DropDown sp Profile Tabs can be split into separate Menu Options
 36. Render Filters on Profile Page using React component
 37. Render SessionSummary on Session Page & Profile Page using Component
+38. Improve Table Handling in SessionSummary Component